Rename slash_config to cog_config & remove default help command
