+++ pg13/cogs/cog_config.py
@@ -7,7 +7,7 @@ config_path = os.environ.get("CONFIG_PATH") or "config.toml"
 config = toml.load(config_path)
 
 # Default to registering commands in all guilds with a provided config
-loaded_guilds = list(map(int, config["guilds"]))
+configured_guilds = list(map(int, config["guilds"]))
 
 # Admin roles/users are configured in config.toml as well
 async def admin_check(interaction: discord.Interaction):
